+++ src/index.ts
@@ -4,6 +4,21 @@ import path from "path";
 
 type BatchRename = (path: string) => void;
 
+const isWSL = async () => {
+  if (os.platform() !== "linux") {
+    return false;
+  }
+
+  try {
+    const procVersion = (
+      await fs.readFile("/proc/version", "utf-8")
+    ).toLowerCase();
+    return procVersion.includes("wsl") && procVersion.includes("microsoft");
+  } catch (err) {
+    return false;
+  }
+};
+
 // Opens system file explorer
 const openFileExplorer = () => {
   const isWsl = process.env.WSL_DISTRO_NAME !== undefined;
@@ -18,9 +33,17 @@ const batchRename: BatchRename = async (pathString) => {
 
 const batchRenameCLI: BatchRename = async (pathString) => {
   const files = await fs.readdir(pathString);
+  console.log("pathString:", pathString);
   console.log("files:", files);
-  const content = await fs.opendir(pathString);
-  console.log("content:", content);
+  // Prompt user if they want to proceed or cancel
+  // Loop through files asynchronously and cancel operation if an error occurs
+
+  // Option 1
+  // Copy file names
+  // Iterate and rename copied file names
+
+  // Option 2
+  // Iterate and rename file names
 };
 
 /* BRAINSTORMING
@@ -31,4 +54,4 @@ const batchRenameCLI: BatchRename = async (pathString) => {
  * 2. User selects a folder they want to rename files
  *  What if user selects files?
  */
-export { batchRename, batchRenameCLI, openFileExplorer };
+export { batchRename, batchRenameCLI, openFileExplorer, isWSL };