+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 ---
+### 18 SEP 2025
+- Defined `mockFsPromises` variable with the value `fs as jest.Mocked<typeof fs>`, which allows `readFile.mockResolvedValue` to be called.
+- Mocked the `fs/promises` module.
+---
+### 17 SEP 2025
+- Created `isWSL` function; returns a boolean depending if the user is running the process from WSL.
+---
 ### 16 SEP 2025
 - The `setupEnv` module will run once per test and checks if `process.env.WSL_WINDOWS_DIR` is defined.
 - Created `.test.env` and `setupEnv` modules in `tests` subdirectory.